+++ Spark_Streaming/pom.xml
@@ -78,7 +78,6 @@
         </dependency>
     </dependencies>
 
-
     <build>
         <resources>
             <resource>
@@ -106,3 +105,4 @@
         </plugins>
     </build>
 </project>
+
