+++ Spark_Streaming/src/main/java/com/hzgc/streaming/alarm/AddAlarmMessage.java
@@ -3,7 +3,6 @@ package com.hzgc.streaming.alarm;
 import java.io.Serializable;
 
 /**
- *
  * 新增告警推送信息类（刘善彬）
  */
 public class AddAlarmMessage implements Serializable {
@@ -25,6 +24,7 @@ public class AddAlarmMessage implements Serializable {
 
     /**
      * 类构造函数
+     *
      * @param alarmType
      * @param dynamicID
      * @param dynamicDeviceID
@@ -34,10 +34,13 @@ public class AddAlarmMessage implements Serializable {
         this.dynamicID = dynamicID;
         this.dynamicDeviceID = dynamicDeviceID;
     }
+
     public AddAlarmMessage() {
     }
 
-    /**Getter and Setter**/
+    /**
+     * Getter and Setter
+     **/
     public String getAlarmType() {
         return alarmType;
     }
