+++ Spark_Streaming/src/main/java/com/hzgc/streaming/alarm/RecognizeAlarmMessage.java
@@ -28,17 +28,22 @@ public class RecognizeAlarmMessage implements Serializable {
      */
     private Item[] items;
 
-    /**构造函数**/
+    /**
+     * 构造函数
+     **/
     public RecognizeAlarmMessage(String alarmType, String dynamicDeviceID, String dynamicID, Item[] items) {
         this.alarmType = alarmType;
         this.dynamicDeviceID = dynamicDeviceID;
         this.dynamicID = dynamicID;
         this.items = items;
     }
+
     public RecognizeAlarmMessage() {
     }
 
-    /**Getter and Setter**/
+    /**
+     * Getter and Setter
+     **/
     public String getAlarmType() {
         return alarmType;
     }
