+++ Spark_Streaming/src/main/resources/sparkJob.properties
@@ -5,14 +5,13 @@ kafka.metadata.broker.list=172.18.18.107:21005,172.18.18.108:21005,172.18.18.109
 kafka.FaceRecognizeAlarmJob.group.id=FaceRecognizeAlarmJobGroupID
 #新增告警读取kafka的topic groupId
 kafka.FaceAddAlarmJob.group.id=FaceAddAlarmJobGroupID
-
-
 #实时任务配置信息
+#spark streaming时间间隔（单位：秒）
+job.recognizeAlarm.timeInterval=2
 job.recognizeAlarm.appName=FaceRecognizeAlarmJob
 job.recognizeAlarm.master=yarn-client
-
+job.addAlarm.timeInterval=2
 job.addAlarm.appName=FaceAddAlarmJob
 job.addAlarm.master=yarn-client
-
 job.offLine.appName=FaceOffLineAlarmJob
 job.offLine.master=yarn-client
